Extend stdlib with native LLVM str_repeat and fix JIT type checks.
Implement str_repeat from scratch in LLVM, fix is_type __value__ tag mapping
and trim index widths, add TYPE_TYPE_ASSERT in JIT, and extend compliance specs.

+++ ext/standard/string_ltrim.php
@@ -60,11 +60,11 @@ final class string_ltrim extends Internal
         $len = $context->builder->load(
             $context->builder->structGep($str, $map['length'])
         );
-        $i32 = $context->getTypeFromString('int32');
-        $zero = $i32->constInt(0, false);
+        $i64 = $context->getTypeFromString('int64');
+        $zero = $i64->constInt(0, false);
         $charPtr = $context->builder->structGep($str, $map['value']);
 
-        $startSlot = $context->builder->alloca($i32, 1, 'ltrim_start');
+        $startSlot = $context->builder->alloca($i64, 1, 'ltrim_start');
         $context->builder->store($zero, $startSlot);
         string_trim::advanceWhileTrimByte($context, $charPtr, $len, $startSlot, true);
